test(toggle): migrate legacy basic
